models/requirements.txt
flask
flask-cors
numpy
pandas
scikit-learn
joblib
google.generativeai
